change the code by PR's feedback
