+++ src/mappings/gvault.ts
@@ -4,12 +4,12 @@ import {
 import {
 	updateGTokenFactor as updateFactor
 } from '../setters/factors';
-import { setGvtPrice } from "../setters/price" 
+import { setGvtPrice } from '../setters/price';
 
 export function handleStrategyHarvestReport(event: LogStrategyHarvestReport): void {
     // update factor
-    updateFactor(event.block.timestamp.toI32())
+    updateFactor(event.block.timestamp.toI32());
 
     // update gvt price
-    setGvtPrice()
+    setGvtPrice();
   }
