+++ src/setters/price.ts
@@ -5,7 +5,6 @@ import {
     getPricePerShare,
     tokenToDecimal
 } from '../utils/tokens';
-import { get3CrvVirtualPrice } from "../utils/threePool";
 import {
     log,
     BigInt,
@@ -25,7 +24,7 @@ import { Vault as BalancerGroWethVault } from '../../generated/BalancerGroWethVa
 import { WeightedPool as BalancerGroWethPool } from '../../generated/BalancerGroWethPool/WeightedPool';
 import { Vyper_contract as CurveMetapool3CRV } from '../../generated/CurveMetapool3CRV/Vyper_contract';
 import { AccessControlledOffchainAggregator as ChainlinkAggregator } from '../../generated/ChainlinkAggregator/AccessControlledOffchainAggregator';
-
+import { ThreePool } from '../../generated/ChainlinkAggregator/ThreePool';
 
 export const initPrice = (): Price => {
     let price = Price.load('0x');
@@ -55,9 +54,17 @@ export const setGvtPrice = (): void => {
 }
 
 export const set3CrvPrice = (): void => {
-    let price = initPrice();
-    price.threeCrv = get3CrvVirtualPrice();
-    price.save();
+    const contract = ThreePool.bind(ADDR.THREE_POOL);
+    const virtualPrice = contract.try_get_virtual_price();
+    if (virtualPrice.reverted) {
+        log.error('Get virtual price for 3crv failed', []);
+    } else {
+        const crvPrice = tokenToDecimal(virtualPrice.value, 18, 7);
+        let price = initPrice();
+        price.threeCrv = crvPrice;
+        price.save();
+    }
+    
 }
 
 // TODO: update gro price as well (knowing gvt price, we can update gro)
